# ewm_main/src/main/java/ru/practicum/main_server/dto/user/UserShortDto.java
package ru.practicum.main_server.dto.user;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.validation.constraints.NotBlank;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class UserShortDto {

    @NotBlank
    long id;

    @NotBlank
    String name;
}